+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to secrets-store-csi-driver-provider-gcp will be documented 
 
 ### Added
 
-* `-sm_connection_pool_size` flag added with default value of `5`. Added as part of an optimization to reuse the same Secret Manager client across all `Mount` operations. [#94](https://github.com/GoogleCloudPlatform/secrets-store-csi-driver-provider-gcp/issues/94)
+* `-sm_connection_pool_size` and `-iam_connection_pool_size` flags added with default value of `5`. Added as part of an optimization to reuse the same Secret Manager client across all `Mount` operations. [#94](https://github.com/GoogleCloudPlatform/secrets-store-csi-driver-provider-gcp/issues/94)
 
 ### Removed